requirements.txt
streamlit==1.31.1
python-dotenv==1.0.1
langchain-core==0.1.24
langchain-community==0.0.21
feedparser==6.0.11
duckduckgo-search==4.4.3
pydantic==2.6.1
requests==2.31.0
plotly==5.18.0
pandas==2.2.0